A first reasonably complete implementation for #309
I have rewritten the way configurations get loaded, so as to avoid the need for promises in the core of the code

In particular it is possible to use the 'rise' key in a notebook metadata in addition to the legacy 'livereveal'  one; 'rise' takes precedence when applicable.

One thing that I have not been able to tackle at all is how to package the new files so that everything works smoothly after installation through pip or conda.

The file configurator.md, as well as rise.png, are accessories that show up in the configurator page, they are just placeholders for now and need to be more elaborate

A few fixes were made to the example slides in the mix, in particular to reflect the use of 'rise' rather than 'livereveal' in the metadata. This probably requires quite a few further changes in the documentation.